requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic
python-dotenv
flet
httpx
Pillow>=10.0.0
